Refactor into ldp package and update docs
