+++ README.md
@@ -1,6 +1,8 @@
 # QXDM Log Analysis Pipeline
 
-This project contains a set of scripts for analysing 3GPP/QXDM log files. The pipeline extracts sentences from large log archives, computes a variety of metrics and supports both Windows and Linux environments.
+This repository contains a small toolkit for processing 3GPP/QXDM log files. It extracts technical sentences from large archives and computes a range of metrics using sentence embeddings.
+
+The code has been refactored into the `ldp` Python package (located under `src/ldp`). You can run the pipeline either directly or via `python -m ldp`.
 
 ## Features
 
@@ -10,21 +12,43 @@ This project contains a set of scripts for analysing 3GPP/QXDM log files. The pi
 - Metrics including semantic spread, redundancy index, cluster entropy and more
 - Optional integration with Weights & Biases for experiment tracking
 
-## Quick Start
-
-Install the required packages and run the main pipeline script:
+## Installation
 
 ```bash
 pip install -r requirements.txt
-python src/Main.py --help
 ```
 
-Example logs for testing can be found in `test_logs/`.
+Optionally install the package in editable mode:
+
+```bash
+pip install -e .
+```
+
+## Usage
+
+Run the pipeline with:
+
+```bash
+python -m ldp --help
+```
+
+Important environment variables:
+
+- `QXDM_ROOT` – directory containing your log files
+- `BLOCK_SIZE` – number of characters processed per block (defaults adaptively)
+- `BATCH_SIZE` – embedding batch size
+- `OMP_NUM_THREADS` – CPU worker count
+
+Example logs for testing can be found in `test_logs/` and `test_dataset/`.
+
+## Repository layout
+
+```
+src/ldp/            Core package modules
+notebooks/          Example Jupyter notebooks
+scripts/            Helper shell scripts
+```
 
 ## License
 
 This repository is provided for research and experimentation purposes. See the individual source files for license details.
-
-## Dataset
-
-The test_dataset folder is used for the test code
