+++ pyproject.toml
@@ -1,0 +1,29 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ldp"
+version = "0.1.0"
+description = "Tools for analysing QXDM/3GPP log files"
+authors = [{name = "QXDM Team"}]
+readme = "README.md"
+requires-python = ">=3.9"
+license = {text = "MIT"}
+dependencies = [
+    "numpy",
+    "pandas",
+    "torch",
+    "sentence-transformers",
+    "scikit-learn",
+    "scipy",
+    "psutil",
+    "tqdm",
+    "spacy",
+    "pdfminer.six",
+    "pynvml",
+    "wandb",
+]
+
+[project.scripts]
+ldp = "ldp.main:main"
