+++ src/__init__.py
@@ -1,0 +1,3 @@
+"""Compatibility wrapper for the old `src` package."""
+
+from .ldp import *  # re-export for backward compatibility
