+++ src/ldp/__init__.py
@@ -1,0 +1,21 @@
+"""London data processing utilities."""
+
+from .main import main
+from .log_processor import QXDMLogProcessor
+from .metric_utils import (
+    semantic_spread,
+    redundancy_index,
+    cluster_entropy,
+    change_mag,
+    novelty_density,
+)
+
+__all__ = [
+    "main",
+    "QXDMLogProcessor",
+    "semantic_spread",
+    "redundancy_index",
+    "cluster_entropy",
+    "change_mag",
+    "novelty_density",
+]
