+++ src/ldp/log_processor.py
